requirements.txt
# Core dependency for .docx processing
python-docx>=1.1.0
lxml>=4.9.0
